requirements.txt
# Core ML & DL
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP
nltk>=3.8.0
regex>=2023.6.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.14.0
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
jsonlines>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
